update for most recent three go versions (#537)
