+++ samlsp/fetch_metadata_test.go
@@ -1,6 +1,7 @@
 package samlsp
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"net/http"
@@ -26,4 +27,22 @@ func TestFetchMetadata(t *testing.T) {
 	md, err := FetchMetadata(context.Background(), testServer.Client(), *u)
 	assert.Check(t, err)
 	assert.Check(t, is.Equal("https://idp.testshib.org/idp/shibboleth", md.EntityID))
+}
+
+func TestFetchMetadataRejectsInvalid(t *testing.T) {
+	test := NewMiddlewareTest(t)
+	test.IDPMetadata = bytes.ReplaceAll(test.IDPMetadata,
+		[]byte("<EntityDescriptor "), []byte("<EntityDescriptor ::foo=\"bar\">]]"))
+
+	testServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		assert.Check(t, is.Equal("/metadata", r.URL.String()))
+		_, err := w.Write(test.IDPMetadata)
+		assert.Check(t, err)
+	}))
+
+	fmt.Println(testServer.URL + "/metadata")
+	u, _ := url.Parse(testServer.URL + "/metadata")
+	md, err := FetchMetadata(context.Background(), testServer.Client(), *u)
+	assert.Check(t, err != nil)
+	assert.Check(t, is.Nil(md))
 }
